Phase 2.3: eunomia std stdio (debug-log) + exit/abort terminus
Wire stdout/stderr (and panic last-words) to the kernel DebugWrite
debug-log path (rev2§7's EL0 scaffold) and override the PAL exit()/
abort_internal() to the thread_exit terminus, so a std binary's
println! is observable and its panic/OOM/exit reaps correctly.

- sys/stdio/eunomia.rs (new): separate Stdin/Stdout/Stderr structs
  (motor shape). Stdout/Stderr delegate to the eunomia-sys seam
  __eunomia_stdio_write (which chunks <=1024 bytes — the kernel
  ERR_FAULTs a longer DebugWrite). Stdin is EOF until the 5.1 console;
  panic_output uses a dedicated PanicWriter so 5.1 keeps panic
  last-words on debug-log while moving stdout/stderr to the console.
- sys/stdio/mod.rs: cfg_select eunomia arm.
- sys/exit.rs: eunomia arm -> __eunomia_thread_exit(code as u32 as u64).
- sys/pal/eunomia/common.rs: abort_internal -> thread_exit(u64::MAX ==
  STATUS_PANIC), replacing intrinsics::abort() so panic/OOM reap as the
  reserved status (rev2§5.1), distinct from exit(0).
- sys/pal/eunomia/mod.rs: _start zero-extends in lockstep so exit(-1)
  cannot collide with the all-ones STATUS_PANIC sentinel.

The EL0 debug-log use of stdout/stderr is a disclosed, temporary
deviation from rev2§2 (rev2§2.7), replaced by the console channel in
std-port 5.1; only panic last-words stay on the path.

+++ library/std/src/sys/pal/eunomia/mod.rs
@@ -39,6 +39,9 @@ pub extern "C" fn _start() -> ! {
 
     // Orderly exit: the parent reaper reads this status. A panic in a std binary instead
     // routes through `abort_internal` → `thread_exit(STATUS_PANIC)` (overridden in
-    // std-port 2.3), which the reaper distinguishes from `exit(0)`.
-    unsafe { __eunomia_thread_exit(code as u64) }
+    // std-port 2.3), which the reaper distinguishes from `exit(0)`. Zero-extend the
+    // `i32` (`code as u32 as u64`, in lockstep with the `sys::exit::exit` eunomia arm):
+    // sign-extending `-1` would land on `u64::MAX == STATUS_PANIC` and reap a clean
+    // `main`-returns-`-1` as a crash.
+    unsafe { __eunomia_thread_exit(code as u32 as u64) }
 }
